voodoo fix for what appears to be validation failure
blkperl observes this when running with master/agent

    Error: /Stage[main]/Mcollective::Common::Config/File[/usr/local/share/mcollective]: Could not evaluate: Could not retrieve information from environment production source(s)
    Notice: /Stage[main]/Mcollective::Common::Config/Datacat_collector[mcollective::site_libdir]: Dependency File[/usr/local/share/mcollective] has failures: true
    Warning: /Stage[main]/Mcollective::Common::Config/Datacat_collector[mcollective::site_libdir]:
    Skipping because of failed dependencies

Which is odd, as the datacat_collector has the ordering relationship of
*before* that file resource.

What I guess is happening is the file resource is failing some form
of early validation, but only when in master/agent mode as this has been
repeatedly tested with puppet apply.

There is always the danger that this means datacat isn't collecting structures
correctly when used master/agent, and so I need to test that more carefully
soon.